Remove 'interpretation(UnobservedActions, current_timestep)'
There is no point in this rule since the controller is not allowed
to assert an interpretation/2 such that this constraint would be violated.
It's sole purpose is to differentiate between Step 1 and Step 2 of the AIA control loop.
We already do this using ASP subprograms, so interpretation/2 is redundant.

Furthermore, the UnobservedActions value of interpretation/2 is never used
(even in Blount's implementation) so interpretation/2 can be removed completely.